requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
gunicorn==21.2.0
